+++ R/curve_analytic.R
@@ -195,8 +195,7 @@ curve_analytic <- function(estimate, se = NULL, df = NULL, n = NULL,
   class(densdf) <- c("data.frame", "concurve")
 
   if (isTRUE(table)) {
-    levels <- c(0.25, 0.50, 0.75, 0.80, 0.85, 0.90, 0.95, 0.975, 0.99)
-    df_subintervals <- curve_table(df_out, levels, type = "c", format = "data.frame")
+    df_subintervals <- curve_table(df_out, type = "c", format = "data.frame")
     class(df_subintervals) <- c("data.frame", "concurve")
     dataframes <- list(df_out, densdf, df_subintervals)
     names(dataframes) <- c("Intervals Dataframe", "Intervals Density", "Intervals Table")