+++ R/curve_corr.R
@@ -74,8 +74,7 @@ curve_corr <- function(x, y, alternative, method, steps = 10000, cores = getOpti
 
 
   if (table == TRUE) {
-    levels <- c(0.25, 0.50, 0.75, 0.80, 0.85, 0.90, 0.95, 0.975, 0.99)
-    (df_subintervals <- (curve_table(df, levels, type = "c", format = "data.frame")))
+    (df_subintervals <- (curve_table(df, type = "c", format = "data.frame")))
     class(df_subintervals) <- c("data.frame", "concurve")
     dataframes <- list(df, densdf, df_subintervals)
     names(dataframes) <- c("Intervals Dataframe", "Intervals Density", "Intervals Table")
@@ -87,4 +86,3 @@ curve_corr <- function(x, y, alternative, method, steps = 10000, cores = getOpti
 }
 
 # RMD Check
-utils::globalVariables(c("df", "lower.limit", "upper.limit", "intrvl.width", "intrvl.level", "cdf", "pvalue", "svalue"))