+++ R/curve_likelihood.R
@@ -79,8 +79,7 @@ as_curve_lik <- function(values, loglik, table = TRUE) {
   class(likfunction) <- c("data.frame", "concurve")
 
   if (isTRUE(table)) {
-    levels <- c(0.03, 0.05, 0.12, 0.14)
-    df_subintervals <- curve_table(likfunction, levels, type = "l", format = "data.frame")
+    df_subintervals <- curve_table(likfunction, type = "l", format = "data.frame")
     class(df_subintervals) <- c("data.frame", "concurve")
     dataframes <- list(likfunction, df_subintervals)
     names(dataframes) <- c("Intervals Dataframe", "Intervals Table")