+++ R/curve_lmer.R
@@ -78,8 +78,7 @@ curve_lmer <- function(object, parm, method = "profile", zeta = NULL,
 
 
   if (table == TRUE) {
-    levels <- c(0.25, 0.50, 0.75, 0.80, 0.85, 0.90, 0.95, 0.975, 0.99)
-    (df_subintervals <- (curve_table(df, levels, type = "c", format = "data.frame")))
+    (df_subintervals <- (curve_table(df, type = "c", format = "data.frame")))
     class(df_subintervals) <- c("data.frame", "concurve")
     dataframes <- list(df, densdf, df_subintervals)
     names(dataframes) <- c("Intervals Dataframe", "Intervals Density", "Intervals Table")
@@ -91,4 +90,3 @@ curve_lmer <- function(object, parm, method = "profile", zeta = NULL,
 }
 
 # RMD Check
-utils::globalVariables(c("df", "lower.limit", "confint.merMod", "upper.limit", "intrvl.width", "intrvl.level", "cdf", "pvalue", "svalue"))