+++ R/curve_wrap.R
@@ -136,8 +136,7 @@ curve_wrap <- function(ci_func, steps = 1000, cores = getOption("mc.cores", 1L),
 
   # Build output list
   if (isTRUE(table)) {
-    levels <- c(0.25, 0.50, 0.75, 0.80, 0.85, 0.90, 0.95, 0.975, 0.99)
-    df_subintervals <- curve_table(df, levels, type = "c", format = "data.frame")
+    df_subintervals <- curve_table(df, type = "c", format = "data.frame")
     class(df_subintervals) <- c("data.frame", "concurve")
 
     dataframes <- list(df, densdf, df_subintervals)
@@ -248,9 +247,3 @@ curve_model <- function(model, param, method = "default", steps = 1000,
   curve_wrap(ci_func, steps = steps, cores = cores, table = table)
 }
 
-
-# R CMD check
-utils::globalVariables(c(
-  "lower.limit", "upper.limit", "intrvl.width",
-  "intrvl.level", "cdf", "pvalue", "svalue"
-))